refactor: extract async step handling, centralize context serialization, and move nested reactor test helpers to support files
