+++ spec/nested_reactor_inline_execution_spec.rb
@@ -3,27 +3,6 @@
 require "spec_helper"
 
 RSpec.describe "Nested Reactor Inline Execution" do
-  class NestedInlineChildReactor < RubyReactor::Reactor
-    input :id
-
-    step :async_step do
-      async true
-      run { |args, _| RubyReactor::Success("async_done_#{args[:id]}") }
-    end
-  end
-
-  class NestedInlineRootReactor < RubyReactor::Reactor
-    input :id
-
-    step :prepare do
-      run { |_, _| RubyReactor::Success("prepared") }
-    end
-
-    compose :child_process, NestedInlineChildReactor do
-      argument :id, input(:id)
-    end
-  end
-
   it "correctly merges state when running inline with nested reactors" do
     # Mock async_router to run inline and return the executor
     allow(RubyReactor.configuration.async_router).to receive(:perform_async) do |serialized_context, reactor_class_name|
@@ -32,7 +11,7 @@ RSpec.describe "Nested Reactor Inline Execution" do
       worker.perform(serialized_context, reactor_class_name)
     end
 
-    result = NestedInlineRootReactor.run(id: "123")
+    result = Support::NestedInlineRootReactor.run(id: "123")
 
     expect(result).to be_a(RubyReactor::Success)
     # The result of the root reactor is a hash of all step results
